refactor(cli): remove leftover inline reporting vars; use shared.reporting
